feat: Add second inventory for edubadges
As a devops, when I run the playbook with ./run, I want to avoid running
it on the wrong POC server. So that I don't undo work from colleagues or
break servers.

With a default, when we don't provide an environment, it will run on the default. We
probably don't ever want that. So removing the default and requiring the
ENV to be provided. Either as arg or as ENV var:

It can be set as arg, so ./run eduwallet or ./run opebadges
Or as ENV var: export INV=eduwallet.