feat: Replace the additional hosts for agent with additional tags
This gives more freedom and is simpler.

Use case is that at OB PoC we want to serve did files from the agent
instead of a third party service. For that, we need to configure traefik
so it can add a path prefix and listen on an extra host. And request a
ssl cert for that.

We can now do that with an override:

```
  agent:
    extra_docker_compose_labels:
      # Add an extra route that proxies to the agent for our did domain.
      # See under 'dids'. Append a did name to the path.
      # The agent then serves did files at:
      #   https://agent.example.com/foobar/.well-known/did.json
      # But the did:web expects it at:
      #   https://foobar.example.com/.well-known/did.json
      - "traefik.http.routers.edubadges.rule=Host(`edubadges.{{ deploydomain }}`)"
      - "traefik.http.routers.edubadges.entrypoints=websecure"
      - "traefik.http.routers.edubadges.tls.certresolver=myresolver"
      - "traefik.http.middlewares.edubadges-addprefix.addprefix.prefix=/edubadges"
      - "traefik.http.routers.edubadges.middlewares=edubadges-addprefix"
```